daml-stdlib: align HasField instance with ghc
For some reason we ended up with a `setField: a -> r -> r` while ghc has
a `setField: r -> a -> r`. This is a source of confusion and subtile
bugs. This PR aligns our `setField` method with the one of ghc in
preparation of the coming nested record update feature.

CHANGELOG_BEGIN
[daml-stdlib]: Switched type of the `setField : a -> r -> r`  method of
the `HasField` class to `setField : r -> a -> r`. The HasField instances
are automatically generated, hence no action needs to be taken to cope
with this change.
CHANGELOG_END