Fix validation of rollback nodes (#9479)
* Fix validation of rollback nodes

To my own embarassment there is no test in this PR. We don’t have any
direct tests for validation atm. We absolutely need some and I already
discussed with Remy that this is something we should focus on once
exceptions are done but for now I’ll stick to being sad.

changelog_begin
changelog_end

* Fix bug

changelog_begin
changelog_end