Backport ContractKeysValidation fixes from #9478
Given that this is a backport, I tried to stay somewhat close to #9478
rather than staying close to the current state. There are a bunch of
simplifications we could make if we just want to minimize the diff
against the current state but that seems more confusing than helpful
imho.

changelog_begin
changelog_end