Only archive a key if it was brought into scope before
This PR implements the semantics we agreed on in
https://github.com/digital-asset/daml/pull/9472#discussion_r620097844.

Before, archive always marked the key as inactive in `keys`. This is
weirdly inconsistent with other operations. For example, a regular
fetch does not bring a key in scope in `keys`. This PR changes this to
a more consistent model where `keys` is modified exactly under the
following circumstances:

1. A create with a key always overwrites.
2. An archive will only mark a key as inactive if there was an entry
   in `keys` with the same contract id.
3. Lookup/fetch by key first check if we currently have an entry in `keys`.
   If so, we use that.
   If not, we need to ask the ledger. We first check if we have a
   ceched entry in `globalKeyInputs`. If we do use that.
   If we don’t, we ask the ledger.
   If the result is not active or not visible we proceed updating `keys`
   and `globalKeyInputs`.
   If the result was a cid, we check if it’s still active.

changelog_begin
changelog_end