Share code between SBUFetchKey and SBULookupKey (#9537)
* Share code between SBUFetchKey and SBULookupKey

Those do almost the same and the logic is growing increasingly complex
so I really want to share it. This also makes it easier to see the
things that are different.

Along the way I’ve also discovered a bug:
SBUFetchKey called crash where it should produce a proper error that
can be caught by submitMustFail. Not going to resolve that in this PR
though so I just added a TODO.

changelog_begin
changelog_end

* review comments

changelog_begin
changelog_end