Cache global key inputs (#9538)
* Cache global key inputs

This PR changes speedy such that it will query at most once for a global
key. This is a prerequisite for a (sensible) duplicate key detection
and for sensible multi-key semantics.

Note that there is one outstanding issue here:

As a general rule, we only want by-key operations to bring a key in
scope not just a random operation like an exercise or a fetch that
happens to have a key.

However, for consuming exercises this is currently not the case. They
always bring the key in scope. The plan here is for an archive to not
block a further lookup and instead add an activeness check to the
lookup afterwards.

changelog_begin
changelog_end

* Update daml-lf/interpreter/src/main/scala/com/digitalasset/daml/lf/speedy/SBuiltin.scala

Co-authored-by: Remy <remy.haemmerle@daml.com>

* s/CtrlImportValue/SEImportValue/

changelog_begin
changelog_end

* .

changelog_begin
changelog_end

* Use a custom Scala type instead of Option[Value.ContractId]

changelog_begin
changelog_end

Co-authored-by: Remy <remy.haemmerle@daml.com>