Throw on internal errors instead of setting ptx to aborted (#9654)
* Throw on internal errors instead of setting ptx to aborted

The main purpose of aborted is to be able to get out a ptx in the
scenario service. However, for internal errors that makes no
sense. Users should never encounter them. And the use of `aborted`
here has swallowed these errors silently in a few places before which
is clearly bad.

changelog_begin
changelog_end

* Fix tryHandleSubmitMustFail

changelog_begin
changelog_end