Add a ledger model page for exceptions. (#9396)
* Add a ledger model page for exceptions.

This page describes the changes to the ledger model as part of the exceptions feature. The changes should be kept separate for now, but once exceptions are stable and out for a while, they should probably be incorporated into the rest of the daml ledger model.

Part of #8020. **This PR is to be merged only when we start releasing exceptions.**

changelog_begin

- [docs] The Daml ledger model has been updated to describe the changes introduced by the new exceptions feature. See  here: https://docs.daml.com/concepts/ledger-model/ledger-exceptions.html

changelog_end

* Apply suggestions from code review

Co-authored-by: Ognjen Maric <ognjen.maric@digitalasset.com>

* Round 2

* Fix and simplify "after" relation!

* Easy fixes

* Diagrams for structure and integrity sections.

* Add missing diagrams :)

* clarify consistency section

* Authorization.

* Editing.

* fix label reference

Co-authored-by: Ognjen Maric <ognjen.maric@digitalasset.com>