Make sure we don’t start leaking directories in compat tests (#9759)
As evidenced by this change, we don’t actually need the safeguard
here. i think this was just copied over from the main workspace where
we did leak installations. I’d like to know when we start leaking
installations here in the future so this PR removes the catch.

changelog_begin
changelog_end