Fix the notion of transaction equivalence in the ledger model for exceptions.
@andreaslochbihler-da found that our normalization rules don't normalize rollbacks across exercises, which our notion of transaction equivalence expects them to. For example, the transactions

```
rollback [ exercise [ rollback [ create ]]]
```

and

```
rollback [ exercise [ create ] ]
```

are considered equivalent under the old definition, but do not normalize to each other.

This PR is one possible solution, where we keep our current normalization rules but change the definition of transaction equivalence. In particular, instead of just requiring that the set of rolled back actions to be the same, we require that the set of "rollback roots" be the same, where a "rollback root" is an action that has a rollback node as its parent. This prevents normalization across exercises because that would eliminate some rollback roots.

The idea here is that the set of "rollback roots" encodes the set of "rolled back actions" relative to each action's subtree, unlike before where we only looked at the set of "rolled back actions" relative to the whole transaction. The set of "rollback roots" encodes all that information, but with a simple definition.

A completely different solution is to introduce new normalization rules that work across exercise nodes. Unfortunately these rules would be non-local, and that seems very difficult to do correctly & efficiently.

Part of #8020

changelog_begin
changelog_end