update Daml Hub auth in create-daml-app (#9973)
* update Daml Hub auth in create-daml-app

Current documentation says the token is passed as a cookie. This is a
little bit more secure than passing it as a URL param, as it used to be.
Note that while it is no longer documented behaviour, the token is
actually still passed as a param (in addition to the cookie), so
existing users are not broken (yet). Still, better if the code we
generate for them matches our current docs.

CHANGELOG_BEGIN
CHANGELOG_END

* review comment