[Scala client bindings] Fix a "deadlock" in CommandRetryFlow
The "deadlock" occurs when a command is sent to the RETRY_PORT, but at
the same time new commands are coming in as well. If the
commandSubmissionFlow is limited (e.g. via the MaxInFlight flow), then
backpressure is exerted on the merge flow and ultimately back to the
retryDecider. This in turn prevents new command results from being
propagated to the PROPAGATE_PORT, and therefore the limited
commandSubmissionFlow ends up backpressuring itself into a "deadlock".

To circumvent this, two new measures are in place:
1. The Merge flow is now a MergePreferred flow, giving precedence to
submission retries.
2. An additional buffer is added to the retry feedback loop, so that the
limited commandSubmissionFlow can keep making progress.

Ideally this buffer would have an additional feature of rerouting new
elements to a completely different port if the buffer is full, but that
would likely require a custom graph stage.

CHANGELOG_BEGIN
CHANGELOG_END