participant-integration-api: Move transaction requests to structured logging. [KVL-996] (#10141)
* participant-integration-api: Remove the subscription ID.

Pretty sure it's not used except for a single log line, which makes it
useless as a correlation ID.

If we want to correlate logs, let's add a correlation ID.

* participant-integration-api: Move transaction requests to trace logging.

Most of the useful information is already in the logging context. We
don't need to log the data structure too.

CHANGELOG_BEGIN
- [Ledger API Server] The amount of data logged in the API transaction
  service has been reduced at INFO level. Please enable TRACE logging to
  log the request data structures.
CHANGELOG_END

* participant-integration-api: Reorder methods in ApiTransactionService.

* participant-integration-api: Add the word "request" to some log lines.

* participant-integration-api: Add a logging prefix for string offsets.

* ledger-api-domain: `immutable.Set` -> `Set`.

It's an alias.

* participant-integration-api: Log transaction filters on subscription.

* participant-integration-api: Log transaction filters.

Just the parties isn't enough information.

* participant-integration-api: Log the entire transaction request.

Structured, because otherwise it's hard to throw things away later.

* contextualized-logging: Avoid `View` because it's not in Scala 2.12.

* contextualized-logging: Add tests for booleans.

* contextualized-logging: Avoid methods that accept views.

Scala 2.12 really doesn't like me when I do that.

* participant-integration-api: One more try at building with Scala 2.12.