Mark Java Bot tests flaky (#10322)
See https://github.com/digital-asset/daml/issues/10273 for details on
the issue.

Summary is that the logic in the Java bots is fundamentally broken
since it modifies the pending set asynchronously so it can trigger for
the same contract multiple times.

While we can certainly fix that or hack around the bug in the tests,
given that the Java bots are deprecated, marking it flaky seems like a
better use of our time.

changelog_begin
changelog_end