fix besu listing
The current entry has been moved from "open-source integrations" to
"commercial integrations" in 73b38f8add. However, the move was done
verbatim, while the format for the two tables is different, making the
Besu listing stick out a bit. Moreover, since the listing was added the
URL of the press release has changed, landing people on the front page
of the DA blog (from which it is possible to find the relevant press
release, but quite frankly googling for it is faster).

For the record, I still think this page should be hosted elsewhere as
the docs release cycle doesn't offer enough dynamism for it.

CHANGELOG_BEGIN
CHANGELOG_END