remove duplicated rule in create-daml-app dlint (#10352)
When opening #10342 I was only aware of the `.gitignore`; @cocreature's
comment made me find out the `.dlint.yaml`, and then, testing it, I
realized we ended up with the ignore rule twice. It does not seem to be
causing any issue in cursory tests, but still, I think it's worth
cleaning up.

CHANGELOG_BEGIN
CHANGELOG_END