participant-integration-api: Fill out stubs in ApiSubmissionServiceSpec. (#10349)
* participant-integration-api: Fill out stubs in ApiSubmissionServiceSpec.

This test fails sometimes and it's hard to figure out why because of the
number of benign stack traces in the logs, caused by missing stubs.

By filling out the stubs so that `LedgerConfigProvider` can use them, we
get much clearer test logs, which should hopefully help us diagnose the
error properly next time the test fails in CI.

CHANGELOG_BEGIN
CHANGELOG_END

* participant-integration-api: Fix flake in ApiSubmissionServiceSpec.

Writing the configuration is non-deterministic and unnecessary.

* participant-integration-api: Use resources in ApiSubmissionServiceSpec.

Not always the best idea to use a resource after it's been released. And
this isn't Rust, so we can't force that at compile time.