add triggers to gsg
Our team has recently had two new joiners, both of which were asked to
develop a small Daml application as part of their onboarding, and both
of which ended up with a great use-case for triggers while not knowing
they existed. I also needed to look into how triggers work for unrelated
reasons, so I thought it would make sense to add triggers to the GSG.

This is of course not meant to cover all of triggers (we already have
doc pages for that), but just to make sure new Daml users know triggers
exist.

CHANGELOG_BEGIN
CHANGELOG_END