participant-integration-api: Split LedgerConfigProvider into two, and add test cases. [KVL-1046] (#10455)
* participant-integration-api: Split `LedgerConfigProvider` into two.

We now have a read-only `CurrentLedgerConfiguration`, and a write-only
`LedgerConfigProvisioner`.

CHANGELOG_BEGIN
CHANGELOG_END

* participant-integration-api: Improve logging in LedgerConfigProvisioner.

Making use of the logging context.

* participant-integration-api: Extract a CurrentLedgerConfiguration trait.

This highlighted a few issues in the tests where we were unnecessarily
calling methods.

* participant-integration-api: Split LedgerConfigProviderSpec.

We can test the two components individually now.

* participant-integration-api: Add a test for streaming the config.

* participant-integration-api: Add tests for provisioning configuration.

* participant-integration-api: Move configuration code into a new package.

* participant-integration-api: Fake ledger config in tests.

Now that there's an interface, we can use it without worrying about
streaming from the index.

* participant-integration-api: Use `List`, not `Seq`, in test fixtures.

Scala 2.12 thinks `Seq.apply` yields a possibly-mutable sequence, which
is incompatible with `Source.apply`.

* participant-integration-api: Explain how `LedgerConfigProvider` works.